clean up services, models and move to shared/models folders
