+++ src/app/dashboard/account-overview/energy-use-heat-map/energy-use-heat-map.component.ts
@@ -5,9 +5,11 @@ import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
 import { IdbFacility, IdbUtilityMeter } from 'src/app/models/idb';
 import * as _ from 'lodash';
-import { HeatMapData, VisualizationService } from 'src/app/utility/visualization/visualization.service';
+import { VisualizationService } from 'src/app/utility/visualization/visualization.service';
 import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
-import { DashboardService, FacilitySummary } from '../../dashboard.service';
+import { DashboardService } from '../../dashboard.service';
+import { FacilitySummary } from 'src/app/models/dashboard';
+import { HeatMapData } from 'src/app/models/visualization';
 @Component({
   selector: 'app-energy-use-heat-map',
   templateUrl: './energy-use-heat-map.component.html',
