+++ src/app/dashboard/facility-overview/facility-heat-map/facility-heat-map.component.ts
@@ -5,7 +5,8 @@ import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
 import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
 import { IdbFacility, IdbUtilityMeter } from 'src/app/models/idb';
-import { HeatMapData, VisualizationService } from '../../../utility/visualization/visualization.service';
+import { HeatMapData } from 'src/app/models/visualization';
+import { VisualizationService } from '../../../utility/visualization/visualization.service';
 
 @Component({
   selector: 'app-facility-heat-map',
