+++ src/app/utility/calanderization/calanderization.component.html
@@ -5,9 +5,9 @@
 
     <div class="controls">
         <!--Column Filters-->
-        <div class="filter item-left">
+        <!-- <div class="filter item-left">
             <button class="btn btn-small btn-solid">Filter</button>
-        </div>
+        </div> -->
         <!--Items Per Page-->
         <div class="items-per-page item-right">
             <span>Show </span>
