+++ src/app/utility/visualization/visualization.service.ts
@@ -1,7 +1,9 @@
 import { Injectable } from '@angular/core';
 import { IdbUtilityMeter } from 'src/app/models/idb';
-import { CalanderizationService, CalanderizedMeter, MonthlyData } from '../calanderization/calanderization.service';
+import { CalanderizationService } from '../../shared/helper-services/calanderization.service';
 import * as _ from 'lodash';
+import { CalanderizedMeter, MonthlyData } from 'src/app/models/calanderization';
+import { HeatMapData } from 'src/app/models/visualization';
 @Injectable({
   providedIn: 'root'
 })
@@ -11,7 +13,7 @@ export class VisualizationService {
 
   getFacilityBarChartData(meters: Array<IdbUtilityMeter>, sumByMonth: boolean, removeIncompleteYears: boolean): Array<{ time: string, energyUse: number, energyCost: number }> {
     //calanderize meters
-    let calanderizedMeterData: Array<CalanderizedMeter> = this.calanderizationService.calanderizeMetersInFacility(meters, true);
+    let calanderizedMeterData: Array<CalanderizedMeter> = this.calanderizationService.getCalanderizedMeterData(meters, true, true);
 
     //create array of just the meter data
     let combindedCalanderizedMeterData: Array<MonthlyData> = calanderizedMeterData.flatMap(meterData => {
@@ -87,13 +89,7 @@ export class VisualizationService {
   getMeterHeatMapData(meters: Array<IdbUtilityMeter>, facilityName: string, removeIncompleteYears: boolean, inAccount: boolean): HeatMapData {
     let months: Array<string> = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
     //calanderize meters
-    let calanderizedMeterData: Array<CalanderizedMeter>;
-    if (inAccount) {
-      calanderizedMeterData = this.calanderizationService.calanderizeMetersInAccount(meters, false);
-    } else {
-      calanderizedMeterData = this.calanderizationService.calanderizeMetersInFacility(meters, false);
-    }
-
+    let calanderizedMeterData: Array<CalanderizedMeter> = this.calanderizationService.getCalanderizedMeterData(meters, inAccount, false);
     //create array of just the meter data
     let combindedCalanderizedMeterData: Array<MonthlyData> = calanderizedMeterData.flatMap(meterData => {
       return meterData.monthlyData;
@@ -142,12 +138,4 @@ export class VisualizationService {
       facilityName: facilityName
     }
   }
-}
-
-
-export interface HeatMapData {
-  resultData: Array<{ monthlyEnergy: Array<number>, monthlyCost: Array<number> }>,
-  months: Array<string>,
-  years: Array<number>,
-  facilityName: string
 }
